pref: divide ai.
